Fix al anular bitacora sin prestamos

+++ php/planilla.php
@@ -87,14 +87,16 @@ $app->post('/anular_bitacora', function () {
 
         if ($d->idplnmovimiento == 3) {
             $finiquito = $db->getQuery("SELECT id, idprestamos, fecha FROM plnfiniquito WHERE idplnempleado = $idempleado AND pendiente = 1")[0];
-            $idprestamos = isset($finiquito->idprestamos) ? explode(',', $finiquito->idprestamos) : null;
+            $idprestamos = $finiquito->idprestamos > 0 ? explode(',', $finiquito->idprestamos) : null;
 
-            foreach ($idprestamos as $id) {
-                $monto = $db->getOneField("SELECT monto FROM plnpresabono WHERE idplnprestamo = $id AND fecha = '$finiquito->fecha'");
-                $db->doQuery("UPDATE plnprestamo SET saldo = $monto, finalizado = 0, liquidacion = null WHERE id = $id");
-                // ver prestamos 
-                $db->doQuery("DELETE FROM plnpresabono WHERE idplnprestamo = $id AND fecha = '$finiquito->fecha'");
-                $db->doQuery("DELETE FROM plnarchivo WHERE DATE_FORMAT(fecha, '%Y-%m-%d') = '$finiquito->fecha' AND idplnarchivotipo = 3 AND idplnempleado = $idempleado");
+            if (isset($idprestamos)) { 
+                foreach ($idprestamos as $id) {
+                    $monto = $db->getOneField("SELECT monto FROM plnpresabono WHERE idplnprestamo = $id AND fecha = '$finiquito->fecha'");
+                    $db->doQuery("UPDATE plnprestamo SET saldo = $monto, finalizado = 0, liquidacion = null WHERE id = $id");
+                    // ver prestamos 
+                    $db->doQuery("DELETE FROM plnpresabono WHERE idplnprestamo = $id AND fecha = '$finiquito->fecha'");
+                    $db->doQuery("DELETE FROM plnarchivo WHERE DATE_FORMAT(fecha, '%Y-%m-%d') = '$finiquito->fecha' AND idplnarchivotipo = 3 AND idplnempleado = $idempleado");
+                }
             }
 
             $db->doQuery("DELETE FROM plnfiniquito WHERE id = $finiquito->id");
